_ams_import/requirements.txt
Django>=5.0,<6.0
django-fsm-2>=3.0
django-allauth>=0.63
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-htmx>=1.17
whitenoise>=6.6
Pillow>=10.0

# Production / Docker
gunicorn>=21.0
mysqlclient>=2.2.0
dj-database-url>=2.0
python-decouple>=3.8
cryptography>=42.0
